Create specific AWS component class and decouple from main form AWS specifics
