+++ ui/admin/app/components/form/storage-bucket/index.hbs
@@ -14,10 +14,6 @@
     pluginTypes=this.pluginTypes
     changePluginType=@changePluginType
     changeCredentialType=@changeCredentialType
-    credentials=this.credentials
-    selectedCredentialType=this.selectedCredentialType
-    updateTypeSelection=this.updateTypeSelection
-    showDynamicCredentials=this.showDynamicCredentials
     rollbackSecretAttrs=this.rollbackSecretAttrs
     toggleDisableCredentialRotation=this.toggleDisableCredentialRotation
   }}
