+++ ui/admin/app/components/form/storage-bucket/index.js
@@ -10,7 +10,6 @@ import { inject as service } from '@ember/service';
 import { set } from '@ember/object';
 
 import {
-  TYPES_CREDENTIALS,
   TYPE_CREDENTIAL_DYNAMIC,
   TYPES_STORAGE_BUCKET_PLUGIN,
 } from 'api/models/storage-bucket';
@@ -23,19 +22,6 @@ export default class FormStorageBucketComponent extends Component {
     'resources.storage-bucket.form.scope.help',
   );
 
-  @tracked selectedCredentialType = this.args.model.credentialType;
-
-  @tracked showDynamicCredentials =
-    this.args.model.credentialType === TYPE_CREDENTIAL_DYNAMIC;
-
-  /**
-   * returns an array of available credential types
-   * @type {array}
-   */
-  get credentials() {
-    return TYPES_CREDENTIALS;
-  }
-
   /**
    * returns an array of available storage bucket plugin types
    */
@@ -44,21 +30,6 @@ export default class FormStorageBucketComponent extends Component {
   }
 
   // =actions
-  /**
-   * Allows to update the credential type
-   * @param type {string}
-   */
-  @action
-  updateTypeSelection(type) {
-    this.selectedCredentialType = type;
-    if (type === TYPE_CREDENTIAL_DYNAMIC) {
-      this.showDynamicCredentials = true;
-    } else {
-      this.showDynamicCredentials = false;
-    }
-    this.args.changeCredentialType(type);
-  }
-
   updateScopeFieldDescription() {
     if (this.args.model.scopeModel.isGlobal) {
       this.scopeFieldDescription = this.intl.t(
